TRAN-193 Add pendingMessageId() fun in Transfer entity

+++ data/src/main/java/mega/privacy/android/data/mapper/transfer/TransferAppDataMapper.kt
@@ -35,7 +35,7 @@ internal class TransferAppDataMapper @Inject constructor() {
                     VoiceClip -> TransferAppData.VoiceClip
                     CameraUpload -> TransferAppData.CameraUpload
                     ChatUpload -> values.firstIfNotBlank()
-                        ?.let { TransferAppData.ChatUpload(it) }
+                        ?.let { TransferAppData.ChatUpload(it.toLong()) }
 
                     SDCardDownload -> {
                         values.firstIfNotBlank()?.let {
